mcp_server/mcp/requirements.txt
playwright>=1.49.0
python-dotenv==1.0.0
gspread>=6.0.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

